+++ src/main/java/com/tristankechlo/toolleveling/client/screen/ItemValueScreen.java
@@ -2,6 +2,7 @@ package com.tristankechlo.toolleveling.client.screen;
 
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.tristankechlo.toolleveling.client.screen.widgets.ItemValuesListWidget;
+import com.tristankechlo.toolleveling.config.ItemValueConfig;
 import com.tristankechlo.toolleveling.config.ToolLevelingConfig;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.Component;
@@ -33,7 +34,7 @@ public class ItemValueScreen extends Screen {
         this.itemValues.render(matrixStack, mouseX, mouseY, partialTicks); // render item list widget
         super.render(matrixStack, mouseX, mouseY, partialTicks); // render buttons
         drawCenteredString(matrixStack, font, Component.translatable("screen.toolleveling.default_item_value_worth",
-                ToolLevelingConfig.defaultItemWorth.getValue()), width / 2, 10, 0xFFFFFF);
+                ItemValueConfig.defaultItemWorth.getValue()), width / 2, 10, 0xFFFFFF);
     }
 
     @Override