+++ src/main/java/com/tristankechlo/toolleveling/client/screen/widgets/ItemValuesListWidget.java
@@ -1,7 +1,7 @@
 package com.tristankechlo.toolleveling.client.screen.widgets;
 
 import com.tristankechlo.toolleveling.client.screen.ItemValueScreen;
-import com.tristankechlo.toolleveling.config.ItemValues;
+import com.tristankechlo.toolleveling.config.ItemValueConfig;
 import net.minecraft.client.gui.components.ObjectSelectionList;
 import net.minecraft.core.NonNullList;
 import net.minecraft.util.Tuple;
@@ -35,7 +35,7 @@ public class ItemValuesListWidget extends ObjectSelectionList<ItemValueEntry> {
     }
 
     private void refreshList() {
-        List<Tuple<ItemStack, Long>> entries = ItemValues.itemValues.entrySet().stream()
+        List<Tuple<ItemStack, Long>> entries = ItemValueConfig.itemValues.getValue().entrySet().stream()
                 .sorted((o1, o2) -> o2.getValue().compareTo(o1.getValue()))
                 .map((entry) -> new Tuple<>(new ItemStack(entry.getKey()), entry.getValue()))
                 .collect(Collectors.toList());