+++ src/main/java/com/tristankechlo/toolleveling/config/primitives/BooleanValue.java
@@ -1,7 +1,9 @@
-package com.tristankechlo.toolleveling.config.values;
+package com.tristankechlo.toolleveling.config.primitives;
 
+import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.tristankechlo.toolleveling.ToolLeveling;
+import com.tristankechlo.toolleveling.config.util.AbstractConfigValue;
 
 public final class BooleanValue extends AbstractConfigValue<Boolean> {
 
@@ -32,7 +34,10 @@ public final class BooleanValue extends AbstractConfigValue<Boolean> {
     @Override
     public void deserialize(JsonObject jsonObject) {
         try {
-            value = jsonObject.get(getIdentifier()).getAsBoolean();
+            JsonElement element = jsonObject.get(getIdentifier());
+            if (element != null) {
+                value = element.getAsBoolean();
+            }
         } catch (Exception e) {
             value = defaultValue;
             ToolLeveling.LOGGER.warn("Error while loading the config value " + getIdentifier() + ", using defaultvalue instead");